Add rest api to get Genre and Game data

+++ src/main/java/hpbui/gamerportal/repository/AccountRepository.java
@@ -1,10 +1,9 @@
 package hpbui.gamerportal.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import hpbui.gamerportal.entity.Account;
 
-public interface AccountRepository extends JpaRepository<Account, Long>{
+public interface AccountRepository extends JpaRepository<Account, Integer>{
 	Account findByEmail(String email);
 }
